requirements.txt
viam-sdk==0.43.1
python-periphery